+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+### New Features
++ Added support for table data queries.
+
 ### Improvements
 + Fixed incorrect default query settings.
 + Refactored query function to use pipes instead of nested observables.